Move show_fps() to utils/display.py
