+++ utils/display.py
@@ -19,6 +19,17 @@ def show_help_text(img, help_text):
                 (32, 32, 32), 4, cv2.LINE_AA)
     cv2.putText(img, help_text, (10, 20), cv2.FONT_HERSHEY_PLAIN, 1.0,
                 (240, 240, 240), 1, cv2.LINE_AA)
+    return img
+
+
+def show_fps(img, fps):
+    """Draw fps number at top-left corner of the image."""
+    font = cv2.FONT_HERSHEY_PLAIN
+    line = cv2.LINE_AA
+    fps_text = 'FPS: {:.2f}'.format(fps)
+    cv2.putText(img, fps_text, (11, 20), font, 1.0, (32, 32, 32), 4, line)
+    cv2.putText(img, fps_text, (10, 20), font, 1.0, (240, 240, 240), 1, line)
+    return img
 
 
 def set_display(window_name, full_scrn):
